feat: update font styles to use Orbitron Variable font
The font styles in the component have been updated to use the Orbitron Variable font instead of Space Grotesk. This change was made to improve the visual appearance of the text. Additionally, the "@fontsource-variable/orbitron" package has been added as a dependency in the package.json file to ensure the font is available for use.